faq_chatbot_example-master/faq_chatbot_example-master/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.6.0
boto==2.49.0
boto3==1.12.14
botocore==1.15.14
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
djangorestframework==3.11.0
docutils==0.15.2
gensim==3.8.1
idna==2.9
jmespath==0.9.5
JPype1==0.7.2
konlpy==0.5.2
lxml==4.5.0
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.9.0
sqlparse==0.3.1
tweepy==3.8.0
urllib3==1.25.8
